requirements.txt
pytesseract
pillow
pyautogui
numpy
mss
pywin32
keyboard